# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.AlbumDTO:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      published:
        type: boolean
      release_date:
        type: string
    type: object
...
  github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.CreateAlbumDTO:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
...
  github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.LoginDTO:
    properties:
      name:
        type: string
      password:
        type: string
    required:
    - name
    - password
    type: object
  github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.LoginResponseDTO:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.RefreshDTO:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  internal_adapters_delivery_api.RestErrorBadRequest:
    properties:
      error:
        example: bad request
        type: string
      status:
        example: 400
        type: integer
      timestamp:
        example: "2020-11-10T23:00:00+00:00"
        type: string
    type: object
...
  internal_adapters_delivery_api.RestErrorForbidden:
    properties:
      error:
        example: forbidden
        type: string
      status:
        example: 403
        type: integer
      timestamp:
        example: "2020-11-10T23:00:00+00:00"
        type: string
    type: object
  internal_adapters_delivery_api.RestErrorInternalError:
    properties:
      error:
        example: internal server error
        type: string
      status:
        example: 500
        type: integer
      timestamp:
        example: "2020-11-10T23:00:00+00:00"
        type: string
    type: object
  internal_adapters_delivery_api.RestErrorNotFound:
    properties:
      error:
        example: not found
        type: string
      status:
        example: 404
        type: integer
      timestamp:
        example: "2020-11-10T23:00:00+00:00"
        type: string
    type: object
  internal_adapters_delivery_api.RestErrorUnauthorized:
    properties:
      error:
        example: unauthorized
        type: string
      status:
        example: 401
        type: integer
      timestamp:
        example: "2020-11-10T23:00:00+00:00"
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a Sigma Music API.
  termsOfService: http://swagger.io/terms/
  title: Sigma Music API
  version: "1.0"
paths:
  /albums:
    get:
      consumes:
      - application/json
      description: get all albums
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.AlbumDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorInternalError'
      summary: GetAllAlbums
      tags:
      - album
  /albums/{id}:
    get:
      consumes:
      - application/json
      description: get album by id
      parameters:
      - description: album id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.AlbumDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorInternalError'
      summary: GetAlbumByID
      tags:
      - album
    patch:
      consumes:
      - application/json
      description: publish album
      parameters:
      - description: album id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorUnauthorized'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorForbidden'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorInternalError'
      security:
      - ApiKeyAuth: []
      summary: PublishAlbum
      tags:
      - album
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: login
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.LoginResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorUnauthorized'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorInternalError'
      summary: LogIn
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorUnauthorized'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorInternalError'
      security:
      - ApiKeyAuth: []
      summary: Logout
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: refresh
      parameters:
      - description: refresh payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.RefreshDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.LoginResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorUnauthorized'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorInternalError'
      security:
      - ApiKeyAuth: []
      summary: TokenRefresh
      tags:
      - auth
...
  /musicians/{id}/albums:
    get:
      consumes:
      - application/json
      description: get album by musician id
      parameters:
      - description: musician id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorInternalError'
      summary: GetAlbumByMusicianID
      tags:
      - album
...
  /musicians/{musician_id}/albums:
    post:
      consumes:
      - application/json
      description: create album
      parameters:
      - description: musician id
        in: path
        name: musician_id
        required: true
        type: string
      - description: create album info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.CreateAlbumDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.AlbumDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorUnauthorized'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorForbidden'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorInternalError'
      security:
      - ApiKeyAuth: []
      summary: CreateAlbum
      tags:
      - album
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"